backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async HTTP
httpx==0.26.0

# Database
asyncpg==0.29.0
psycopg[binary]==3.2.3
pgvector==0.3.5
alembic==1.13.1

# Auth & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
cryptography==42.0.0

# Redis
redis==5.0.1
rq==1.16.1

# Code parsing
tree-sitter==0.20.4
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.4
tree-sitter-typescript==0.21.2
tree-sitter-java==0.21.0

# Azure
azure-storage-blob==12.19.0
azure-ai-formrecognizer==3.3.2

# OpenAI
openai==1.10.0
tiktoken==0.6.0

# Utilities
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.2.3

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==24.2.0
isort==5.13.2
ruff==0.2.2
mypy==1.8.0

# Monitoring (optional)
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.0

feedparser==6.0.11
beautifulsoup4==4.12.3
apscheduler==3.10.4
pypdf==4.0.1
python-jose
email-validator

langchain>=0.1.0
langchain-openai>=0.0.8
langchain-community>=0.0.21
langgraph>=0.0.30
